Move node to own file.
