+++ crates/action-runner/src/lib.rs
@@ -1,7 +1,9 @@
 mod dep_graph;
 mod errors;
+mod node;
 mod runner;
 
 pub use dep_graph::*;
 pub use errors::*;
+pub use node::Node;
 pub use runner::*;
